change ah flip price to long

+++ Models/PriceEstimate.cs
@@ -4,9 +4,9 @@ namespace Coflnet.Sky.Sniper.Models
     {
         public ReferencePrice Lbin { get; set; }
         public ReferencePrice SLbin { get; set; }
-        public int Median { get; set; }
+        public long Median { get; set; }
         public float Volume { get; set; }
-        public int FourthPercentile { get; set; }
+        public long FourthPercentile { get; set; }
         public string LbinKey { get; set; }
         public string MedianKey { get; set; }
     }
